requirements.txt
just_playback
ffmpeg-python
pycaw
yt_dlp
PySide6
pydub
tqdm
superqt
mutagen